# pkg/proto/openapi/identity/api/v1/identity.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "IdentityApi",
    "version": "1.0"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/identity/v1/get-identity-updates": {
      "post": {
        "summary": "Used to check for changes related to members of a group.\nWould return an array of any new installations associated with the wallet\naddress, and any revocations that have happened.",
        "operationId": "IdentityApi_GetIdentityUpdates",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/xmtpidentityapiv1GetIdentityUpdatesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/xmtpidentityapiv1GetIdentityUpdatesRequest"
            }
          }
        ],
        "tags": [
          "IdentityApi"
        ]
      }
    },
    "/identity/v1/get-inbox-ids": {
      "post": {
        "summary": "Retrieve the XIDs for the given addresses",
        "operationId": "IdentityApi_GetInboxIds",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiv1GetInboxIdsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiv1GetInboxIdsRequest"
            }
          }
        ],
        "tags": [
          "IdentityApi"
        ]
      }
    },
    "/identity/v1/publish-identity-update": {
      "post": {
        "summary": "Publishes an identity update for an XID or wallet. An identity update may\nconsist of multiple identity actions that have been batch signed.",
        "operationId": "IdentityApi_PublishIdentityUpdate",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1PublishIdentityUpdateResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1PublishIdentityUpdateRequest"
            }
          }
        ],
        "tags": [
          "IdentityApi"
        ]
      }
    },
    "/identity/v1/subscribe-association-changes": {
      "post": {
        "operationId": "IdentityApi_SubscribeAssociationChanges",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/v1SubscribeAssociationChangesResponse"
                },
                "error": {
                  "$ref": "#/definitions/rpcStatus"
                }
              },
              "title": "Stream result of v1SubscribeAssociationChangesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1SubscribeAssociationChangesRequest"
            }
          }
        ],
        "tags": [
          "IdentityApi"
        ]
      }
    },
    "/identity/v1/verify-smart-contract-wallet-signatures": {
      "post": {
        "summary": "Verify an unverified smart contract wallet signature",
        "operationId": "IdentityApi_VerifySmartContractWalletSignatures",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1VerifySmartContractWalletSignaturesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1VerifySmartContractWalletSignaturesRequest"
            }
          }
        ],
        "tags": [
          "IdentityApi"
        ]
      }
    }
  },
  "definitions": {
    "GetIdentityUpdatesResponseIdentityUpdateLog": {
      "type": "object",
      "properties": {
        "sequenceId": {
          "type": "string",
          "format": "uint64"
        },
        "serverTimestampNs": {
          "type": "string",
          "format": "uint64"
        },
        "update": {
          "$ref": "#/definitions/associationsIdentityUpdate"
        }
      },
      "description": "A single entry in the XID log on the server."
    },
...
    "SubscribeAssociationChangesResponseAccountAddressAssociation": {
      "type": "object",
      "properties": {
        "accountAddress": {
          "type": "string"
        },
        "inboxId": {
          "type": "string"
        }
      },
      "title": "A change in account_address -\u003e inbox_id association"
    },
    "SubscribeAssociationChangesResponseAccountAddressRevocation": {
      "type": "object",
      "properties": {
        "accountAddress": {
          "type": "string"
        },
        "inboxId": {
          "type": "string"
        }
      },
      "title": "A revocation of an account_address -\u003e inbox_id association"
    },
    "apiv1GetInboxIdsRequest": {
      "type": "object",
      "properties": {
        "requests": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/apiv1GetInboxIdsRequestRequest"
          }
        }
      },
      "title": "Request to retrieve the XIDs for the given addresses"
    },
    "apiv1GetInboxIdsRequestRequest": {
      "type": "object",
      "properties": {
        "address": {
          "type": "string"
        }
      },
      "title": "A single request for a given address"
    },
    "apiv1GetInboxIdsResponse": {
      "type": "object",
      "properties": {
        "responses": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/apiv1GetInboxIdsResponseResponse"
          }
        }
      },
      "title": "Response with the XIDs for the requested addresses"
    },
    "apiv1GetInboxIdsResponseResponse": {
      "type": "object",
      "properties": {
        "address": {
          "type": "string"
        },
        "inboxId": {
          "type": "string"
        }
      },
      "title": "A single response for a given address"
    },
...
    "associationsCreateInbox": {
      "type": "object",
      "properties": {
        "initialAddress": {
          "type": "string"
        },
        "nonce": {
          "type": "string",
          "format": "uint64"
        },
        "initialAddressSignature": {
          "$ref": "#/definitions/identityassociationsSignature",
          "title": "Must be an addressable member"
        }
      },
      "description": "The first entry of any XID log. The XID must be deterministically derivable\nfrom the address and nonce.\nThe recovery address defaults to the initial associated_address unless\nthere is a subsequent ChangeRecoveryAddress in the log."
    },
    "associationsIdentityAction": {
      "type": "object",
      "properties": {
        "createInbox": {
          "$ref": "#/definitions/associationsCreateInbox"
        },
        "add": {
          "$ref": "#/definitions/associationsAddAssociation"
        },
        "revoke": {
          "$ref": "#/definitions/associationsRevokeAssociation"
        },
        "changeRecoveryAddress": {
          "$ref": "#/definitions/associationsChangeRecoveryAddress"
        }
      },
      "title": "A single identity operation"
    },
    "associationsIdentityUpdate": {
      "type": "object",
      "properties": {
        "actions": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/associationsIdentityAction"
          }
        },
        "clientTimestampNs": {
          "type": "string",
          "format": "uint64"
        },
        "inboxId": {
          "type": "string"
        }
      },
      "description": "One or more identity actions that were signed together.\nExample: [CreateXid, AddAssociation, ChangeRecoveryAddress]\n1. The batched signature text is created by concatenating the signature text\n   of each association together with a separator, '\\n\\n\\n'.\n2. The user signs this concatenated result.\n3. The resulting signature is added to each association proto where relevant.\n   The same signature may be used for multiple associations in the array."
    },
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1GetIdentityUpdatesRequestRequest": {
      "type": "object",
      "properties": {
        "inboxId": {
          "type": "string"
        },
        "sequenceId": {
          "type": "string",
          "format": "uint64"
        }
      },
      "description": "Points to the last entry the client has received. The sequence_id should be\nset to 0 if the client has not received anything."
    },
    "v1GetIdentityUpdatesResponseResponse": {
      "type": "object",
      "properties": {
        "inboxId": {
          "type": "string"
        },
        "updates": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/GetIdentityUpdatesResponseIdentityUpdateLog"
          }
        }
      },
      "title": "The update log for a single identity, starting after the last cursor"
    },
    "v1PublishIdentityUpdateRequest": {
      "type": "object",
      "properties": {
        "identityUpdate": {
          "$ref": "#/definitions/associationsIdentityUpdate"
        }
      },
      "title": "Publishes an identity update to the network"
    },
    "v1PublishIdentityUpdateResponse": {
      "type": "object",
      "title": "The response when an identity update is published"
    },
    "v1SubscribeAssociationChangesRequest": {
      "type": "object",
      "title": "Request to subscribe to association changes, triggered by Identity Updates"
    },
    "v1SubscribeAssociationChangesResponse": {
      "type": "object",
      "properties": {
        "timestampNs": {
          "type": "string",
          "format": "uint64"
        },
        "accountAddressAssociation": {
          "$ref": "#/definitions/SubscribeAssociationChangesResponseAccountAddressAssociation"
        },
        "accountAddressRevocation": {
          "$ref": "#/definitions/SubscribeAssociationChangesResponseAccountAddressRevocation"
        }
      }
    },
    "v1VerifySmartContractWalletSignatureRequestSignature": {
      "type": "object",
      "properties": {
        "accountId": {
          "type": "string",
          "title": "CAIP-10 string\nhttps://github.com/ChainAgnostic/CAIPs/blob/main/CAIPs/caip-10.md"
        },
        "blockNumber": {
          "type": "string",
          "format": "uint64",
          "title": "Specify the block number to verify the signature against"
        },
        "signature": {
          "type": "string",
          "format": "byte",
          "title": "The signature bytes"
        },
        "hash": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "v1VerifySmartContractWalletSignaturesRequest": {
      "type": "object",
      "properties": {
        "signatures": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1VerifySmartContractWalletSignatureRequestSignature"
          }
        }
      }
    },
    "v1VerifySmartContractWalletSignaturesResponse": {
      "type": "object",
      "properties": {
        "responses": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1VerifySmartContractWalletSignaturesResponseValidationResponse"
          }
        }
      }
    },
    "v1VerifySmartContractWalletSignaturesResponseValidationResponse": {
      "type": "object",
      "properties": {
        "isValid": {
          "type": "boolean"
        },
        "blockNumber": {
          "type": "string",
          "format": "uint64"
        },
        "error": {
          "type": "string"
        }
      }
    },
    "xmtpidentityapiv1GetIdentityUpdatesRequest": {
      "type": "object",
      "properties": {
        "requests": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1GetIdentityUpdatesRequestRequest"
          }
        }
      },
      "title": "Get all updates for an identity since the specified time"
    },
    "xmtpidentityapiv1GetIdentityUpdatesResponse": {
      "type": "object",
      "properties": {
        "responses": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1GetIdentityUpdatesResponseResponse"
          }
        }
      },
      "title": "Returns all log entries for the requested identities"
    },
...
  }
}